pyproject.toml
[project]
name = "housegallery"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-slugify==8.0.4",
    "Pillow==11.0.0",
    "argon2-cffi==23.1.0",
    "uvicorn[standard]==0.32.1",
    "uvicorn-worker==0.2.0",
    "psycopg2-binary>=2.9.3",
    # Django
    "django==5.0.10",
    "django-environ==0.11.2",
    "django-model-utils==5.0.0",
    "django-crispy-forms==2.3",
    "crispy-bootstrap5==2024.10",
    # Django REST Framework
    "djangorestframework==3.15.2",
    "django-cors-headers==4.6.0",
    "drf-spectacular==0.28.0",
    # Django asset helpers
    "django-webpack-loader==3.1.1",
    "whitenoise>6.0.0",
    "django-storages[google]>1.13.0",
    # Wagtail
    "wagtail==6.3.2",
    # Google Cloud
    "google-api-python-client>2.45.0",
    "google-cloud-logging>3.1.2",
    "google-cloud-secret-manager>2.16.1",
]

[dependency-groups]
dev = [
    "Werkzeug[watchdog]==3.1.3",
    "ipdb==0.13.13",
    "watchfiles==1.0.3",
    # Testing
    "mypy==1.13.0",
    "django-stubs[compatible-mypy]==5.1.1",
    "pytest==8.3.4",
    "pytest-sugar==1.0.0",
    "djangorestframework-stubs==3.15.1",
    # Documentation
    "sphinx==8.1.3",
    "sphinx-autobuild==2024.10.3",
    # Code quality
    "ruff==0.8.3",
    "coverage==7.6.9",
    "djlint==1.36.3",
    "pre-commit==4.0.1",
    # Django
    "factory-boy==3.3.1",
    "django-extensions==3.2.3",
    "django-coverage-plugin==3.1.0",
    "pytest-django==4.9.0",
    "django-querycount==0.8.3",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["housegallery/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true
